Refactor: naming and directory structure for tests
Now, while the tests are still collocated, now the tests are further
split by directory in: unit, integration, system and e2e.
