+++ orders/package.json
@@ -11,7 +11,7 @@
     "preset": "ts-jest",
     "testEnvironment": "node",
     "setupFilesAfterEnv": [
-      "./src/test/setup.ts"
+      "./src/test/integration/setup.ts"
     ]
   },
   "keywords": [],
