--- orders/src/routes/__test__/delete.test.ts
@@ -1,58 +1,0 @@
-import mongoose from 'mongoose';
-import request from 'supertest';
-import { app } from '../../app';
-import { Ticket } from '../../models/ticket';
-import { Order, OrderStatus } from '../../models/order';
-import { natsWrapper } from '../../nats-wrapper';
-
-it('marks an order as cancelled', async () => {
-  // create a ticket with Ticket Model
-  const ticket = Ticket.build({
-    id: new mongoose.Types.ObjectId().toHexString(),
-    title: 'concert',
-    price: 20,
-  });
-  await ticket.save();
-
-  const user = global.signin();
-  // make a request to create an order
-  const { body: order } = await request(app)
-    .post('/api/orders')
-    .set('Cookie', user)
-    .send({ ticketId: ticket.id })
-    .expect(201);
-  // make a request to cancel the order
-  await request(app)
-    .delete(`/api/orders/${order.id}`)
-    .set('Cookie', user)
-    .send()
-    .expect(204);
-  // expectation to make sure the order is cancelled.
-  const updatedOrder = await Order.findById(order.id);
-  expect(updatedOrder!.status).toEqual(OrderStatus.Cancelled);
-});
-
-it('emits an order cancelled event', async () => {
-  const ticket = Ticket.build({
-    id: new mongoose.Types.ObjectId().toHexString(),
-    title: 'concert',
-    price: 20,
-  });
-  await ticket.save();
-
-  const user = global.signin();
-  // make a request to create an order
-  const { body: order } = await request(app)
-    .post('/api/orders')
-    .set('Cookie', user)
-    .send({ ticketId: ticket.id })
-    .expect(201);
-  // make a request to cancel the order
-  await request(app)
-    .delete(`/api/orders/${order.id}`)
-    .set('Cookie', user)
-    .send()
-    .expect(204);
-
-  expect(natsWrapper.client.publish).toHaveBeenCalled();
-});
