+++ orders/src/routes/__tests__/integration/new.route.integration.test.ts
@@ -1,0 +1,74 @@
+import mongoose from 'mongoose';
+import request from 'supertest';
+import { app } from '../../../app';
+import { Ticket } from '../../../models/ticket';
+import { Order, OrderStatus } from '../../../models/order';
+import { natsWrapper } from '../../../nats-wrapper';
+// NOTE: Add the appropriate tests for this service's routes based on
+// the examples in the 'tickets' service.
+
+it('returns an error if the ticket does not exist', async () => {
+  const ticketId = new mongoose.Types.ObjectId();
+  await request(app)
+    .post('/api/orders')
+    .set('Cookie', global.signin())
+    .send({ ticketId })
+    .expect(404);
+});
+
+it('returns an error if the ticket is already reserved', async () => {
+  const ticket = Ticket.build({
+    id: new mongoose.Types.ObjectId().toHexString(),
+    title: 'concert',
+    price: 20,
+  });
+
+  await ticket.save();
+  const order = Order.build({
+    ticket,
+    userId: 'laskdfklsdjf',
+    status: OrderStatus.Created,
+    expiresAt: new Date(), // NOTE: For this test, this date does not matter
+  });
+  await order.save();
+
+  await request(app)
+    .post('/api/orders')
+    .set('Cookie', global.signin())
+    .send({ ticketId: ticket.id })
+    .expect(400);
+});
+
+it('reserves a ticket', async () => {
+  const ticket = Ticket.build({
+    id: new mongoose.Types.ObjectId().toHexString(),
+    title: 'concert',
+    price: 20,
+  });
+
+  await ticket.save();
+
+  await request(app)
+    .post('/api/orders')
+    .set('Cookie', global.signin())
+    .send({ ticketId: ticket.id })
+    .expect(201);
+});
+
+it('emits an order created event', async () => {
+  const ticket = Ticket.build({
+    id: new mongoose.Types.ObjectId().toHexString(),
+    title: 'concert',
+    price: 20,
+  });
+
+  await ticket.save();
+
+  await request(app)
+    .post('/api/orders')
+    .set('Cookie', global.signin())
+    .send({ ticketId: ticket.id })
+    .expect(201);
+
+  expect(natsWrapper.client.publish).toHaveBeenCalled();
+});
