+++ payments/jest.integration.config.js
@@ -1,8 +1,10 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  testMatch: ['<rootDir>/src/**/*.integration.test.ts'],
-  setupFilesAfterEnv: ['<rootDir>/src/test/integration/setup.ts'],
+  testMatch: [
+    '<rootDir>/src/services/**/__tests__/integration/*.service.integration.test.ts',
+  ],
+  setupFilesAfterEnv: ['<rootDir>/src/test/integration/setup.compose.ts'],
   maxWorkers: 1, // Important for deterministic concurrency tests + less flakiness
   testTimeout: 120000,
 };
