+++ payments/package.json
@@ -13,7 +13,7 @@
     "preset": "ts-jest",
     "testEnvironment": "node",
     "setupFilesAfterEnv": [
-      "./src/test/setup.ts"
+      "./src/test/integration/setup.inmemory.ts"
     ]
   },
   "keywords": [],
