--- payments/src/events/listeners/__test__/order-created-listener.test.ts
@@ -1,46 +1,0 @@
-import { OrderCreatedEvent, OrderStatus } from '@rrpereztickets/common';
-import { JsMsg } from 'nats';
-import { natsWrapper } from '../../../nats-wrapper';
-import { OrderCreatedListener } from '../order-created-listener';
-import mongoose from 'mongoose';
-import { Order } from '../../../models/order';
-
-const setup = async () => {
-  const listener = new OrderCreatedListener(natsWrapper.client);
-
-  const data: OrderCreatedEvent['data'] = {
-    id: new mongoose.Types.ObjectId().toHexString(),
-    version: 0,
-    status: OrderStatus.Created,
-    userId: 'notusedhere',
-    expiresAt: 'notusedhere',
-    ticket: {
-      id: 'notusedhere',
-      price: 10,
-    },
-  };
-  // @ts-ignore
-  const msg: JsMsg = {
-    ack: jest.fn(),
-  };
-
-  return { listener, data, msg };
-};
-
-it('replicates the order info', async () => {
-  const { listener, data, msg } = await setup();
-
-  await listener.onMessage(data, msg);
-
-  const order = await Order.findById(data.id);
-
-  expect(order!.price).toEqual(data.ticket.price);
-});
-
-it('acks the message', async () => {
-  const { listener, data, msg } = await setup();
-
-  await listener.onMessage(data, msg);
-
-  expect(msg.ack).toHaveBeenCalled();
-});
