--- payments/src/test/setup.ts
@@ -1,57 +1,0 @@
-import { MongoMemoryServer } from 'mongodb-memory-server';
-import mongoose from 'mongoose';
-import { app } from '../app';
-import request from 'supertest';
-import jwt from 'jsonwebtoken';
-
-jest.mock('../nats-wrapper');
-
-declare global {
-  var signin: (id?: string) => string[];
-}
-
-let mongo: any;
-beforeAll(async () => {
-  process.env.JWT_KEY = 'ASDFDDSFAS';
-  mongo = await MongoMemoryServer.create();
-  const mongoUri = mongo.getUri();
-
-  await mongoose.connect(mongoUri, {});
-});
-
-beforeEach(async () => {
-  jest.clearAllMocks();
-  const collections = await mongoose.connection.db.collections();
-
-  for (let collection of collections) {
-    await collection.deleteMany({});
-  }
-});
-
-afterAll(async () => {
-  if (mongo) {
-    await mongo.stop();
-  }
-  await mongoose.connection.close();
-});
-
-global.signin = (id?: string) => {
-  // Build a JWT payload. {id, email }
-  const payload = {
-    id: id || new mongoose.Types.ObjectId().toHexString(),
-    email: 'test@test.com',
-  };
-  // Create the JWT!
-  const token = jwt.sign(payload, process.env.JWT_KEY!);
-
-  // Build session Object. { jwt: MY_JWT }
-  const session = { jwt: token };
-
-  // Turn that session into JSON
-  const sessionJSON = JSON.stringify(session);
-  // Take JSON and encode it as base 64
-  const b64session = Buffer.from(sessionJSON).toString('base64');
-  // return a string thats the ckie with encoded data
-
-  return [`session=${b64session}`];
-};
