+++ tickets/src/events/listeners/__tests__/integration/order-created.listener.integration.test.ts
@@ -1,0 +1,66 @@
+import { OrderCreatedListener } from '../../order-created-listener';
+import { natsWrapper } from '../../../../nats-wrapper';
+import { Ticket } from '../../../../models/ticket';
+import mongoose from 'mongoose';
+import { OrderStatus } from '@rrpereztickets/common';
+import { JsMsg, JSONCodec } from 'nats';
+import { OrderCreatedEvent } from '@rrpereztickets/common';
+
+const setup = async () => {
+  const listener = new OrderCreatedListener(natsWrapper.client);
+  const ticket = Ticket.build({ title: 'concert', price: 99, userId: 'sdfa' });
+
+  await ticket.save();
+
+  const data: OrderCreatedEvent['data'] = {
+    id: new mongoose.Types.ObjectId().toHexString(),
+    version: 0,
+    status: OrderStatus.Created,
+    userId: 'asdfasdfadsfa',
+    expiresAt: 'asdfdasfsdf',
+    ticket: {
+      id: ticket.id,
+      price: ticket.price,
+    },
+  };
+
+  // @ts-ignore
+  const msg: JsMsg = {
+    ack: jest.fn(),
+  };
+
+  return { listener, ticket, data, msg };
+};
+
+it('sets the orderId of the ticket', async () => {
+  const { listener, data, msg, ticket } = await setup();
+
+  await listener.onMessage(data, msg);
+
+  const updatedTicket = await Ticket.findById(ticket.id);
+
+  expect(updatedTicket!.orderId).toEqual(data.id);
+});
+
+it('acks the message', async () => {
+  const { listener, data, msg, ticket } = await setup();
+  await listener.onMessage(data, msg);
+
+  expect(msg.ack).toHaveBeenCalled();
+});
+
+it('publishes a ticket updated event', async () => {
+  const { listener, data, msg, ticket } = await setup();
+  await listener.onMessage(data, msg);
+  const jc = JSONCodec();
+
+  expect(natsWrapper.client.publish).toHaveBeenCalled();
+
+  const call = (natsWrapper.client.publish as jest.Mock).mock.calls[0];
+  const subject = call[0];
+  const encodedPayload = call[1];
+  const ticketUpdatedData = jc.decode(encodedPayload); // Decoded payload
+
+  // @ts-ignore
+  expect(data.id).toEqual(ticketUpdatedData.orderId);
+});
